IsoPlugin: remove second isoSprite register...
It added a way to put the game object directly into a group, but in
doing so confuses the user (and the IDE) on what is happening. In my
knowledge this is not the way Phaser uses those registers. They seem
to exist in order to make the game object directly handled by the
scene. If needed, the game object should either be:
- added to a group after its creation using the register
- as it does not need to be handled by the scene, created normally
(new ...IsoSprite), then added to a group.